requirements.txt
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.8
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
python-dotenv==0.19.2
pytz==2021.3
schedule==1.1.0
typing-extensions==4.0.1
yarl==1.7.2
